Remove now-unused Store method
